+++ nion/data/xdata_1_0.py
@@ -195,6 +195,9 @@ def conj(data_and_metadata: _DataAndMetadataLike) -> DataAndMetadata.DataAndMeta
 def real_if_close(data_and_metadata: _DataAndMetadataLike, tol: int=100) -> DataAndMetadata.DataAndMetadata:
     return Core.function_array(numpy.real_if_close, data_and_metadata, tol)
 
+def power(data_and_metadata: _DataAndMetadataLike, exponent: float) -> DataAndMetadata.DataAndMetadata:
+    return Core.function_array(numpy.power, data_and_metadata, exponent)
+
 # rgb
 
 def red(data_and_metadata: _DataAndMetadataLike) -> DataAndMetadata.DataAndMetadata:
@@ -268,6 +271,9 @@ def line_profile(data_and_metadata: _DataAndMetadataLike, vector: Core.NormVecto
 
 def invert(data_and_metadata: _DataAndMetadataLike) -> DataAndMetadata.DataAndMetadata:
     return Core.function_invert(data_and_metadata)
+
+def radial_profile(data_and_metadata: _DataAndMetadataLike, center: typing.Optional[Core.NormPointType] = None) -> DataAndMetadata.DataAndMetadata:
+    return Core.function_radial_profile(data_and_metadata, center)
 
 # sequences: registration, shifting, alignment, integrate, trim, insert, concat, extract